\name{memory}
\alias{memory}
\title{Check the total and free memory in the OS}
\usage{
memory()
}
\description{
This is a function that checks the total amount of memory on the system, and also the amount that is not used (free memory). This uses the commands from operating system, and would be useful to check an object with what size can be fitted in memory.
}
\examples{
memory()
}
